[feat][dingo-executor] Add replace into, autoIncrement support long type, quarter function (#1397)

+++ dingo-calcite/src/main/java/io/dingodb/calcite/grammar/ddl/DingoSqlDdlNodes.java
@@ -45,7 +45,7 @@ public class DingoSqlDdlNodes {
         int replica,
         String engine,
         Properties properties,
-        int autoIncrement,
+        long autoIncrement,
         String comment,
         String charset,
         String collate,
